back/django_rest_framework_test/app/views.py
from django.shortcuts import render
from rest_framework import viewsets, filters

from .models import Recipe
from .serializers import RecipeSerializer

class RecipeViewSet(viewsets.ModelViewSet):
    queryset = Recipe.objects.all()
    serializer_class = RecipeSerializer