back/requirements.txt
Django==4.2.13
dj-database-url==1.2.0
djangorestframework==3.15.2
mysqlclient
Pillow==10.4.0